Prevent overflow in continuation history updates
Continuation-history bonuses can exceed 20k on fail lows. Multiplying the bonus by the history weight and consistency multiplier as an int can overflow before the result is scaled, invoking undefined behavior and perturbing move ordering.

Widen the intermediate product to i64, then convert the safely scaled update back to the int expected by StatsEntry.

Bench: 2449838

+++ src/search.cpp
@@ -2054,7 +2054,7 @@ void update_continuation_histories(Stack* ss, Piece pc, Square to, int bonus) {
                 positiveCount++;
 
             int multiplier = CMHCMultipliers[positiveCount];
-            historyEntry << (bonus * weight * multiplier / 131072) + 73 * (i < 2);
+            historyEntry << int(i64(bonus) * weight * multiplier / 131072) + 73 * (i < 2);
         }
     }
 }
